requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
autopep8==1.5.7
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
dj-rest-auth==2.1.7
Django==3.2.4
django-allauth==0.44.0
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
filelock==3.0.12
idna==2.10
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
oauthlib==3.1.1
pipenv==2021.5.29
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.8.3
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.1
stripe==2.58.0
toml==0.10.2
urllib3==1.26.5
virtualenv==20.4.7
virtualenv-clone==0.5.4
wrapt==1.12.1
